% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_diagonal_spline.R
\name{bounding_diagonal_spline}
\alias{bounding_diagonal_spline}
\title{Makes a straight diagonal spline curve}
\usage{
bounding_diagonal_spline(x, degree, f = 1.2)
}
\arguments{
\item{x}{The data points coordinates, as a matrix of row-vectors}

\item{degree}{The degree of the spline}

\item{f}{A coefficient that determines how interior are the points of the computed diagonal. Higher values
correspond with more interior points}
}
\value{
A matrix of row-vectors representing the control points of a spline curve.
}
\description{
\code{bounding_diagonal_spline} makes a diagonal straight spline from the data point bounding box.
This function returns the control points of the spline, as a matrix of row-vectors. It has been used 
as onset for spline fitting routines in some examples, but it should not be standarized - each specific problem
requires careful consideration.
}
\section{Auxiliary functions}{

}

\author{
Máximo Sánchez-Aragón
}
