% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{B-spline evaluation}
\usage{
bspline(x, k = -1, degree = 3, U = bspline_support(degree), native = TRUE)
}
\arguments{
\item{x}{The parameter value (single value).}

\item{k}{The index of the containing interval for x. If it is
equal to -1 the interval is computed.}

\item{U}{The knot support, as a vector.}

\item{native}{If \code{TRUE} Native code will be used instead.}
}
\value{
A vector of B-splines evaluations
}
\description{
\code{bspline} evaluates all B-splines for the single value \code{x}
}
\details{
This function evaluates all B-splines at a single value of the parameter \code{x}
}
\author{
Máximo Sánchez-Aragón
}
