% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_eval.R
\name{bspline_eval}
\alias{bspline_eval}
\title{Spline evaluation}
\usage{
bspline_eval(x, degree = 3, U = bspline_support(degree), native = TRUE)
}
\arguments{
\item{x}{The parameter values, as a numeric vector.}

\item{degree}{The degree of the polynomials that define the spline}

\item{U}{The support, as a numeric vector.}

\item{native}{If \code{TRUE} a native implementation is used instead}
}
\value{
A matrix containing all B-splines evaluations. Evaluations for the same
parameter value are contained in individual rows. Columns represent different evaluations
of the same B-spline at different parameter values.
}
\description{
\code{bspline_eval} evaluates at once the parameter values of vector \code{x}
}
\details{
This function evaluates all B-splines at the values of the vector \code{x}.
It is important to remark that no warning will be thrown when \code{native} equals \code{TRUE}
for parametric values (\code{x}) that have reached the support highest value.
}
\examples{
options(scipen = 5)
par(mgp=c(2.4,1,.4))
par(mar=c(4,5,2.5,2)+0.1) 
par(cex=.9)

# clamped B-spline evaluation (open) --------------------
knot_number <- 10

# the graphical function
test_spline <- function(xseq, user_knots, p, 
                        sup = bspline_support(p, knots=user_knots)  ){
  
  bbasis <- bspline_eval(xseq, degree=p, U=sup, native=TRUE)
  
  for(i in 1:ncol(bbasis)){
    if(i == 1){
      plot(xseq, bbasis[,1], type = "l", main=paste("p = ",p, sep=""),
           ylab = "B(x)", xlab="x", xaxt="n", ylim=c(0,1))      
    }else{
      points(xseq, bbasis[,i], type = "l")
    }
  }
  
  for(i in 1:length(sup))
    lines(c(sup[i],sup[i]),c(0,1), lty="dotted")
  
  #axis(1, at=sup, labels=sup)
  sup.lab <- round(sup, digits=2)
  axis(1, at=sup, labels=FALSE)#sup.lab)
  text(x=sup, y=par()$usr[3]-0.09*(par()$usr[4]-par()$usr[3]),
       labels=sup.lab, srt=45, adj=1, xpd=TRUE)
  
}

# choose the polinomial degree that you wish 
# splines will be evaluated from 0 to p
p <- 11

# the sequence of evaluation
xseq <- seq(from = 0, to = 1, length.out = 200);

# the knots
user_knots <- c(0,sort(runif(knot_number)),1)

# the chart generation
par(mfrow=c(2,3))
for(j in 0:p) test_spline(xseq, user_knots, j)
rm(j)



# periodic B-spline evaluation (closed) ---------------------------------
# the degree 
p <- 11

# the sequence of evaluation
xseq <- seq(from = 0, to = 0.99999, length.out = 200);

# the knots
user_knots <- sort(c(0,sort(runif(knot_number)),1))

par(mfrow=c(2,3))
for(j in 0:p){ 
  sup <- bspline_support(j, knots=user_knots, native=TRUE)
  sup_per <- transform_support_periodic(degree = j, U = sup, native=TRUE)
  test_spline(xseq, NULL, j, sup=sup_per)
}
rm(j)
}
\author{
Máximo Sánchez-Aragón
}
