% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_find_interval_index.R
\name{bspline_find_interval_index}
\alias{bspline_find_interval_index}
\title{Finding the interval index of a knot support}
\usage{
bspline_find_interval_index(x, U, native = TRUE)
}
\arguments{
\item{x}{The parameter value (single value).}

\item{U}{The support, as a vector.}

\item{native}{If \code{TRUE} native code will be used instead (default).}
}
\value{
the interval index, as an integer.
}
\description{
\code{bspline_find_interval_index} finds the interval
index of the provided parameter value.
}
\details{
The evaluation of B-splines at the parameter value \code{x} is bound to the
interval where \code{x} lies within the B-spline support (i.e. the sequence of
ruputure points or knots). This function computes this interval for the single
value of \code{x}.
}
\author{
Máximo Sánchez-Aragón
}
