% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_length.R
\name{bspline_length}
\alias{bspline_length}
\title{Finding the length of a spline curve by a parameter subinterval}
\usage{
bspline_length(c_point, degree = 3, resolution = 100, begin = 0, end = 1, ...)
}
\arguments{
\item{c_point}{the row-vector matrix of control points}

\item{degree}{the spline polynomial degree}

\item{resolution}{The number of chunks used for length computation. This number is applied to a 
parameter domain [0,1] and it is rescaled accordingly with the arguments of \code{begin} and \code{end}}

\item{begin}{the lower bound of the parameter value interval from which the length of the spline curve is computed}

\item{end}{the upper bound of the parameter value interval from which the length of the spline curve is computed}
}
\description{
It computes the length of a spline curve by the extended sum of the distances of a sequence
of interpolated points in the interval of parameter values specified by the user. The sequence of
interpolated points is chosen from an evenly distributed set of parameter values in the subinterval
that which is intended to measure.
}
\author{
Máximo Sánchez-Aragón
}
