% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_parametrize.R
\name{bspline_parametrize}
\alias{bspline_parametrize}
\title{Finding parameter values for a set of interpolation points}
\usage{
bspline_parametrize(interp.points, method = "centripetal", limit = 1)
}
\arguments{
\item{interp.points}{The interpolation points, as a matrix of row-vector positions}

\item{method}{The interpolation method. Currently there are only
three options: 'uniform', 'chordal' or 'centripetal' (default)}

\item{limit}{The upper bound for the knot sequence, which is set to 1 by default.
Lower bound is set to zero in all cases.}
}
\value{
paramater values for the provided curve points
}
\description{
\code{bspline_parametrize} finds the parameter values for the spatial points
provided, i.e. points belonging to the spline curve, by one of several methods.
}
\author{
Máximo Sánchez-Aragón
}
