% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_support.R
\name{bspline_support}
\alias{bspline_support}
\title{Building a clamped knot support}
\usage{
bspline_support(
  degree,
  knots = c(0:5),
  multiplicities = c(degree + 1, rep(1, length(knots) - 2), degree + 1),
  native = FALSE
)
}
\arguments{
\item{degree}{The degree of the polynomial for the current spline.}

\item{knots}{The non-redundant sequence of knots (rupture points) in the parameter space.}

\item{multiplicities}{A vector of integers, with same length as \code{knots}
that specifies the multiplicities at each position of \code{knots}}

\item{native}{If \code{TRUE} native code is used.}
}
\value{
A support that is adequate for clamped spline evaluation
}
\description{
\code{bspline_support} makes a clamped knot support for splines
}
\details{
This function makes a clamped knot support out of an increasing sequence of
knots. The support creates the necessary multiplicities (but not all)
for the provided degree (i.e. the degree of the polynomials that build the spline).
These multiplicities determine at last the maximum curve class for the spline.
}
\author{
Máximo Sánchez-Aragón
}
