% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_turns.R
\name{bspline_turns}
\alias{bspline_turns}
\title{Turns of a spline curve relative to a reference vector}
\usage{
bspline_turns(
  c_point,
  degree = 3,
  tolerance = 1e-09,
  initial.scan.points = nrow(c_point) * 8,
  knots = seq(from = 0, to = 1, length.out = nrow(c_point) - degree + 1),
  verbose = FALSE,
  max.iterations = 100,
  ref_vector = NULL,
  pick.sides = FALSE,
  ...
)
}
\arguments{
\item{c_point}{The current control points, expressed as a matrix of row vectors.}

\item{degree}{The maximum polynomial degree of the spline curve.}

\item{initial.scan.points}{Number of initial points that are scanned.}

\item{knots}{The sequence of knots of the spline curve support.}

\item{verbose}{If \code{TRUE} it shows information while computing.}

\item{max.iterations}{the maximum number of allowed iterations allowed.
If the value is exceeded, the algorithm interprets this as a non-orthogonal
minimum, and discards it with a warning.}

\item{ref_vector}{The reference vector used. If \code{NULL} the tangent vector at 
u = 0 will be used.}

\item{pick.sides}{Picks maximum and minimum values only from minimum candidates 
for u.}

\item{...}{Other parameters passed to interpolate_bspline.}
}
\value{
A \code{data.frame} that includes the parameter of a turn, along 
other data such as the actual dot product, the second derivative and the 
number of NRA iterations.
}
\description{
This function returns a set of parameter values corresponding with the 
points in a spline curve where a change of pi/2 degrees in the tangent relative to a 
reference vector occurs. 

This function is deprecated after the functions \link{bspline_uz_sign}, 
\link{bspline_parity_sign} and \link{bspline_parity_turns}.
}
\author{
Máximo Sánchez-Aragón
}
