% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_angles.R
\name{projection_angles}
\alias{projection_angles}
\title{Projection angles computation.}
\usage{
projection_angles(ap_u, d_point, c_point, degree = 3, unit = "degree", ...)
}
\arguments{
\item{ap_u}{Set of parameter values defining the points the curve}

\item{d_point}{A row-vector matrix with the data points to be projected.}

\item{c_point}{Control points of the spline curve.}

\item{degree}{The degree of the spline curve.}

\item{unit}{The angle unit, i.e. "radian" or "degree"}

\item{...}{Other arguments passed to \code{interpolate_bspline}}
}
\value{
a vector with the computed angles
}
\description{
\code{projection_angles} computes the angles of data point projections onto the assigned footpoints in a
spline curve. These footpoints are not necessarilly optimized.
}
\author{
Máximo Sánchez-Aragón
}
