% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_support_periodic.R
\name{transform_support_periodic}
\alias{transform_support_periodic}
\title{Turning knot support into periodic.}
\usage{
transform_support_periodic(
  degree = 3,
  U = bspline_support(degree, knots = seq(from = 0, to = 1, length.out = 10)),
  native = FALSE
)
}
\arguments{
\item{degree}{The degree of spline polynomials that use this support.}

\item{U}{A clamped support, created by the function \code{bspline_support}.}

\item{native}{If TRUE native code is used (default).}
}
\value{
A periodic support vector.
}
\description{
\code{transform_support_periodic} turns a clamped support into 
a periodic support.
}
\author{
Máximo Sánchez-Aragón
}
